backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
scikit-learn==1.6.1
numpy==2.2.1
joblib==1.4.2
openai==1.58.1
